third_party/WebKit/LayoutTests/fast/css/zoom-body-scroll-expected.txt

PASS document.scrollingElement.scrollHeight is within 0.00001 of 1000
PASS document.scrollingElement.scrollWidth is within 0.00001 of 1000
PASS document.scrollingElement.scrollLeft is within 0.00001 of 40
PASS document.scrollingElement.scrollHeight is within 0.00001 of 1000
PASS document.scrollingElement.scrollWidth is within 0.00001 of 1000
PASS document.scrollingElement.scrollTop is within 0.00001 of 0
PASS document.scrollingElement.scrollLeft is within 0.00001 of 40
PASS document.scrollingElement.scrollTop is within 0.00001 of 100
PASS document.scrollingElement.scrollTop is within 0.00001 of 100
PASS document.scrollingElement.scrollLeft is within 0.00001 of 40
PASS successfullyParsed is true

TEST COMPLETE



Test for rdar://problem/6643103 Unscaled values from body.scrollHeight.
